bug fix : cmd_code is a long int (not a int)

+++ fcrondyn.h
@@ -21,7 +21,7 @@
  *  `LICENSE' that comes with the fcron source distribution.
  */
 
- /* $Id: fcrondyn.h,v 1.3 2003-12-25 22:42:35 thib Exp $ */
+ /* $Id: fcrondyn.h,v 1.4 2005-02-26 15:09:54 thib Exp $ */
 
 #ifndef __FCRONDYN_H__
 #define __FCRONDYN_H__
@@ -41,7 +41,7 @@ typedef struct cmd_list_ent {
     char *cmd_name;
     char *cmd_desc;
     int cmd_numopt;
-    int cmd_code;
+    long int cmd_code;
     int cmd_opt[MAX_NUM_OPT];
     int cmd_default[MAX_NUM_OPT];
 } cmd_list_ent;
